requirements.txt
Pillow==8.1.0
PyPDF2==1.26.0
